requirements.txt
# for caiman
bokeh
coverage
cython
future
h5py
holoviews
ipykernel==4.10
ipython
ipyparallel
jupyter
matplotlib
mypy
nose
numpy
numpydoc
opencv-python
peakutils
pims
psutil
pynwb
pyqtgraph
pyserial
scikit-image
scikit-learn
scipy
seaborn
tensorflow
tifffile
torch
tqdm
yapf